refactor: extract core extension contract projection hooks
